+++ lib/index.js
@@ -1,12 +1,13 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.mdPluginChemistry = exports.mdPluginList = exports.mdPluginTableTabular = exports.mdPluginAnchor = exports.mdPluginTOC = exports.mdPluginText = exports.mdPluginHighlightCode = exports.mdPluginMathJax = exports.mathpixMarkdownPlugin = exports.MathpixMarkdownModel = exports.MathpixMarkdown = exports.MathpixLoader = void 0;
+exports.mdPluginChemistry = exports.mdPluginList = exports.mdPluginTableTabular = exports.mdPluginAnchor = exports.mdPluginTOC = exports.mdPluginText = exports.mdPluginHighlightCode = exports.mdPluginMathJax = exports.initMathpixMarkdown = exports.mathpixMarkdownPlugin = exports.MathpixMarkdownModel = exports.MathpixMarkdown = exports.MathpixLoader = void 0;
 var mathpix_loader_1 = require("./components/mathpix-loader");
 exports.MathpixLoader = mathpix_loader_1.default;
 var mathpix_markdown_1 = require("./components/mathpix-markdown");
 exports.MathpixMarkdown = mathpix_markdown_1.default;
 var mathpix_markdown_plugins_1 = require("./markdown/mathpix-markdown-plugins");
-exports.mathpixMarkdownPlugin = mathpix_markdown_plugins_1.default;
+Object.defineProperty(exports, "mathpixMarkdownPlugin", { enumerable: true, get: function () { return mathpix_markdown_plugins_1.mathpixMarkdownPlugin; } });
+Object.defineProperty(exports, "initMathpixMarkdown", { enumerable: true, get: function () { return mathpix_markdown_plugins_1.initMathpixMarkdown; } });
 var mdPluginConfigured_1 = require("./markdown/mdPluginConfigured");
 Object.defineProperty(exports, "mdPluginMathJax", { enumerable: true, get: function () { return mdPluginConfigured_1.mdPluginMathJax; } });
 Object.defineProperty(exports, "mdPluginHighlightCode", { enumerable: true, get: function () { return mdPluginConfigured_1.mdPluginHighlightCode; } });
