+++ lib/markdown/index.js
@@ -26,7 +26,7 @@ var mdInit = function (options) {
         lineNumbering: lineNumbering,
         htmlSanitize: htmlSanitize
     })
-        .use(mathpix_markdown_plugins_1.default, mmdOptions)
+        .use(mathpix_markdown_plugins_1.mathpixMarkdownPlugin, mmdOptions)
         .use(require('markdown-it-multimd-table'), { enableRowspan: true, enableMultilineRows: true })
         .use(require("markdown-it-footnote"))
         .use(require("markdown-it-sub"))
