+++ lib/markdown/mathpix-markdown-plugins.d.ts
@@ -1,2 +1,3 @@
-declare const mathpixMarkdownPlugin: (md: any, options: any) => void;
-export default mathpixMarkdownPlugin;
+export declare const mathpixMarkdownPlugin: (md: any, options: any) => void;
+export declare const setBaseOptionsMd: (baseOption: any, mmdOptions: any) => void;
+export declare const initMathpixMarkdown: (md: any, callback: any) => any;
