+++ lib/markdown/mdPluginLists.js
@@ -22,10 +22,10 @@ exports.default = (function (md, options) {
     re_level_1.SetDefaultItemizeLevel();
     re_level_1.SetDefaultEnumerateLevel();
     re_level_1.clearItemizeLevelTokens();
-    md.block.ruler.after("list", "Lists", lists_1.Lists, options);
-    if (options.renderElement && (options.renderElement.class === 'li_enumerate' ||
-        options.renderElement.class === 'li_itemize')) {
-        md.block.ruler.before("Lists", "ReRenderListsItem", lists_1.ReRenderListsItem, options);
+    md.block.ruler.after("list", "Lists", lists_1.Lists, md.options);
+    if (md.options.renderElement && (md.options.renderElement.class === 'li_enumerate' ||
+        md.options.renderElement.class === 'li_itemize')) {
+        md.block.ruler.before("Lists", "ReRenderListsItem", lists_1.ReRenderListsItem, md.options);
     }
     md.block.ruler.before("Lists", "ReNewCommand", renewcommand_1.ReNewCommand);
     md.inline.ruler.before('escape', 'list_begin_inline', lists_2.listBeginInline);
