+++ lib/markdown/mdPluginTableTabular.js
@@ -9,14 +9,14 @@ var sub_tabular_1 = require("./md-block-rule/begin-tabular/sub-tabular");
 var sub_math_1 = require("./md-block-rule/begin-tabular/sub-math");
 var parse_error_1 = require("./md-block-rule/parse-error");
 exports.default = (function (md, options) {
-    var width = options.width;
     begin_table_1.ClearTableNumbers();
     begin_table_1.ClearFigureNumbers();
     sub_tabular_1.ClearSubTableLists();
     sub_math_1.ClearSubMathLists();
     parse_error_1.ClearParseError();
     Object.assign(md.options, options);
-    md.block.ruler.after("fence", "BeginTabular", begin_tabular_1.BeginTabular, options);
+    var width = md.options.width;
+    md.block.ruler.after("fence", "BeginTabular", begin_tabular_1.BeginTabular, md.options);
     md.block.ruler.before("BeginTabular", "BeginAlign", begin_align_1.BeginAlign);
     md.block.ruler.before("BeginAlign", "BeginTable", begin_table_1.BeginTable);
     md.inline.ruler.before("escape", "InlineIncludeGraphics", includegraphics_1.InlineIncludeGraphics);
