+++ lib/markdown/mdPluginText.d.ts
@@ -1,4 +1,5 @@
 import { RuleInline } from 'markdown-it';
 export declare const resetCounter: RuleInline;
+export declare const resetTextCounter: RuleInline;
 declare const _default: () => (md: any) => void;
 export default _default;
