+++ lib/markdown/mdPluginText.js
@@ -1,6 +1,6 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.resetCounter = void 0;
+exports.resetTextCounter = exports.resetCounter = void 0;
 var getAbstractTemplate = function (content) { return "<h4 style=\"text-align: center\">Abstract</h4><p style=\"text-indent: 1em\">" + content + "</p>"; };
 var subsectionParentCount = 0;
 var sectionCount = 0;
@@ -10,6 +10,12 @@ var isNewSect = false;
 var isNewSubSection = false;
 exports.resetCounter = function () {
     subsectionParentCount = 0;
+};
+exports.resetTextCounter = function () {
+    subsectionParentCount = 0;
+    sectionCount = 0;
+    subCount = 0;
+    subSubCount = 0;
 };
 var headingSection = function (state, startLine /*, endLine*/) {
     sectionCount = 0;
