+++ lib/mathpix-markdown-model/index.js
@@ -3,10 +3,12 @@ Object.defineProperty(exports, "__esModule", { value: true });
 exports.MathpixMarkdownModel = void 0;
 var check_formula_1 = require("./check-formula");
 var markdown_1 = require("../markdown");
-var styles_1 = require("../mathjax/styles");
-var styles_tabular_1 = require("../mathjax/styles-tabular");
-var styles_fonts_1 = require("../mathjax/styles-fonts");
-var styles_lists_1 = require("../mathjax/styles-lists");
+var styles_1 = require("../styles");
+var styles_container_1 = require("../styles/styles-container");
+var styles_code_1 = require("../styles/styles-code");
+var styles_tabular_1 = require("../styles/styles-tabular");
+var styles_fonts_1 = require("../styles/styles-fonts");
+var styles_lists_1 = require("../styles/styles-lists");
 var mathjax_1 = require("../mathjax");
 var formatSourceHtml = function (text, notTrim) {
     if (notTrim === void 0) { notTrim = false; }
@@ -230,7 +232,7 @@ var MathpixMarkdown_Model = /** @class */ (function () {
                     var style = document.createElement("style");
                     style.setAttribute("id", "Mathpix-styles");
                     var bodyStyles = isResetBodyStyles ? styles_1.resetBodyStyles : '';
-                    style.innerHTML = bodyStyles + styles_1.MathpixStyle(setTextAlignJustify) + styles_1.codeStyles + styles_tabular_1.tabularStyles + styles_lists_1.listsStyles;
+                    style.innerHTML = bodyStyles + styles_1.MathpixStyle(setTextAlignJustify) + styles_code_1.codeStyles + styles_tabular_1.tabularStyles() + styles_lists_1.listsStyles;
                     document.head.appendChild(style);
                 }
                 return true;
@@ -264,18 +266,28 @@ var MathpixMarkdown_Model = /** @class */ (function () {
             }
         };
         this.getMathpixStyleOnly = function () {
-            var style = _this.getMathjaxStyle() + styles_1.MathpixStyle(false) + styles_1.codeStyles + styles_tabular_1.tabularStyles + styles_lists_1.listsStyles;
+            var style = _this.getMathjaxStyle() + styles_1.MathpixStyle(false) + styles_code_1.codeStyles + styles_tabular_1.tabularStyles() + styles_lists_1.listsStyles;
             return style;
         };
         this.getMathpixStyle = function (stylePreview, showToc, tocContainerName) {
             if (stylePreview === void 0) { stylePreview = false; }
             if (showToc === void 0) { showToc = false; }
             if (tocContainerName === void 0) { tocContainerName = 'toc'; }
-            var style = styles_1.ContainerStyle + _this.getMathjaxStyle() + styles_1.MathpixStyle(stylePreview) + styles_1.codeStyles + styles_tabular_1.tabularStyles + styles_lists_1.listsStyles;
+            var style = styles_container_1.ContainerStyle() + _this.getMathjaxStyle() + styles_1.MathpixStyle(stylePreview) + styles_code_1.codeStyles + styles_tabular_1.tabularStyles() + styles_lists_1.listsStyles;
             if (showToc) { }
             return stylePreview
                 ? showToc ? style + styles_1.PreviewStyle + styles_1.TocStyle(tocContainerName) : style + styles_1.PreviewStyle
                 : style;
+        };
+        this.getMathpixMarkdownStyles = function (useColors) {
+            if (useColors === void 0) { useColors = true; }
+            var style = styles_container_1.ContainerStyle(useColors);
+            style += _this.getMathjaxStyle();
+            style += styles_1.MathpixStyle(false, useColors);
+            // style += codeStyles;
+            style += styles_tabular_1.tabularStyles(useColors);
+            style += styles_lists_1.listsStyles;
+            return style;
         };
         this.getMathpixFontsStyle = function () {
             return styles_fonts_1.fontsStyles;
