+++ lib/styles/index.d.ts
@@ -1,0 +1,4 @@
+export declare const resetBodyStyles = "\n  body {\n    margin: 0;\n    line-height: normal;\n  }\n";
+export declare const MathpixStyle: (setTextAlignJustify?: boolean, useColors?: boolean) => string;
+export declare const PreviewStyle = "\n  #preview {\n    font-family: 'CMU Serif', 'Georgia', Helvetica, Arial, sans-serif;\n    font-size: 17px;\n    visibility: visible;\n    word-break: break-word;\n    padding: 2.5em;\n    max-width: 800px;\n    margin: auto;\n    box-sizing: content-box;\n  }\n\n  #preview h1, #preview h2, #preview h3, #preview h4, #preview h5, #preview strong {\n    font-family: 'CMU Serif Bold', 'Georgia', Helvetica, Arial, sans-serif;\n  }\n\n  #preview  i, #preview  em {\n    font-family: 'CMU Serif Italic', 'Georgia', Helvetica, Arial, sans-serif;\n  }\n";
+export declare const TocStyle: (containerName?: string) => string;
