+++ lib/styles/index.js
@@ -1,0 +1,15 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.TocStyle = exports.PreviewStyle = exports.MathpixStyle = exports.resetBodyStyles = void 0;
+exports.resetBodyStyles = "\n  body {\n    margin: 0;\n    line-height: normal;\n  }\n";
+exports.MathpixStyle = function (setTextAlignJustify, useColors) {
+    if (setTextAlignJustify === void 0) { setTextAlignJustify = true; }
+    if (useColors === void 0) { useColors = true; }
+    return "\n    #setText > div {\n        justify-content: inherit;\n        margin-top: 0;\n        margin-bottom: 1em;\n        " + (setTextAlignJustify ? 'text-align: justify;' : '') + "\n    }\n    \n    #setText div:last-child {\n        margin-bottom: 0 !important;\n    }\n\n    #setText > br, #preview-content br {\n        line-height: 1.2;\n    }\n\n    #preview-content > div {\n        margin-top: 0;\n        margin-bottom: 1em;\n        text-align: justify;\n    }\n\n    #preview-content table {\n      margin-bottom: 1em;\n    }\n\n    #setText table {\n      margin-bottom: 1em;\n    }\n\n    mjx-container {\n      overflow-y: visible !important;\n      padding-top: 1px;\n      padding-bottom: 1px;\n    }\n    \n    .math-inline mjx-container {\n        display: inline-block !important;\n        page-break-inside: avoid;\n    }\n    .math-block {\n        align-items: center;\n        min-width: min-content;\n        page-break-after: auto;\n        page-break-inside: avoid;\n        margin-top: 1em;\n        margin-bottom: 1em;\n    }\n    .math-block p {\n        flex-shrink: 1;\n    }\n    .math-block mjx-container {\n        margin: 0 !important;\n    }\n    .math-error {\n        background-color: yellow;\n        color: red;\n    }\n\n    #preview-content svg, #setText svg {\n        min-width: initial !important;\n    }\n\n    #preview-content img, #setText img {\n        max-width: 100%;\n    }\n    \n    #preview-content blockquote,  #setText blockquote {\n        page-break-inside: avoid;\n        " + (useColors ? 'color: #666;' : '') + "\n        margin: 0 0 1em 0;\n        padding-left: 3em;\n        border-left: .5em solid #eee;\n    }\n\n    #preview-content pre, #setText pre {\n        border: 1px solid #ccc;\n        page-break-inside: avoid;\n        padding: 0.5em;\n        " + (useColors ? 'background: #f8f8fa;' : '') + "\n    }\n    .empty {\n        text-align: center;\n        font-size: 18px;\n        padding: 50px 0 !important;\n    }\n\n    #setText table, #preview-content table {\n        display: block; \n        overflow: auto;\n        width: 100%;\n        border-collapse: collapse;\n        page-break-inside: avoid;\n    }\n      \n    #setText table th, #preview-content table th {\n        text-align: center;\n        font-weight: bold;\n    }\n    \n    #setText table td, #preview-content table td,\n    #setText table th, #preview-content table th {\n        border: 1px solid #dfe2e5;\n        padding: 6px 13px;\n    }\n      \n    #setText table tr, #preview-content table tr {\n        " + (useColors ? 'background-color: #fff;' : '') + "\n        border-top: 1px solid " + (useColors ? '#c6cbd1' : 'currentColor') + ";\n    }\n    \n    #setText table tr:nth-child(2n), #preview-content table tr:nth-child(2n) {\n        " + (useColors ? 'background-color: #f6f8fa;' : '') + "\n    }\n\n    \n    #setText .main-title, #setText .author, #preview-content .main-title, #preview-content .author  {\n        text-align: center;\n        margin: 0 auto;\n    }\n    \n    #preview-content .main-title, #setText .main-title {\n        line-height: 1.2;\n        margin-bottom: 1em;\n    }\n\n    #preview-content .author, #setText .author  {\n        display: flex;\n        justify-content: center;\n        flex-wrap: wrap;\n    }\n\n    #preview-content .author p, #setText .author p {\n        min-width: 30%;\n        max-width: 50%;\n        padding: 0 7px;\n    }\n\n    #preview-content .author span, #setText .author span {\n        display: block;\n        text-align: center;\n    }\n\n    #preview-content .section-title, #setText .section-title {\n        margin-top: 1.5em;\n    }\n\n    #preview-content .abstract, #setText .abstract {\n        text-align: justify;\n        margin-bottom: 1em;\n    }\n\n    #preview-content .abstract p, #setText .abstract p {\n        margin-bottom: 0;\n    }\n\n    @media print {\n\n      #preview {\n        font-size: 10pt!important;\n      }\n\n      svg {\n        shape-rendering: crispEdges;\n      }\n\n      .math-block svg, math-inline svg {\n        margin-top: 1px;\n      }\n\n      #preview-content img, #setText img {\n        display: block;\n      }\n      \n      #preview-content .figure_img img, #setText .figure_img img {\n        display: inline;\n      }\n\n      .preview-right {\n        word-break: break-word;\n      }\n\n      #preview-content h1, #setText h1 {\n        page-break-inside: avoid;\n        position: relative;\n        border: 2px solid transparent;\n      }\n  \n      #preview-content h1::after, #setText h1::after {\n        content: \"\";\n        display: block;\n        height: 100px;\n        margin-bottom: -100px;\n        position: relative;\n      }\n  \n      #preview-content h2, #setText h2 {\n        page-break-inside: avoid;\n        position: relative;\n        border: 2px solid transparent;\n      }\n  \n      #preview-content h2::after, #setText h2::after {\n        content: \"\";\n        display: block;\n        height: 100px;\n        margin-bottom: -100px;\n        position: relative;\n      }\n  \n      #preview-content h3, #setText h3 {\n        page-break-inside: avoid;\n        position: relative;\n        border: 2px solid transparent;\n      }\n  \n      #preview-content h3::after, #setText h3::after {\n        content: \"\";\n        display: block;\n        height: 100px;\n        margin-bottom: -100px;\n        position: relative;\n      }\n  \n      #preview-content h4, #setText h4 {\n        page-break-inside: avoid;\n        position: relative;\n        border: 2px solid transparent;\n      }\n  \n      #preview-content h4::after, #setText h4::after {\n        content: \"\";\n        display: block;\n        height: 100px;\n        margin-bottom: -100px;\n        position: relative;\n      }\n  \n      #preview-content h5, #setText h5 {\n        page-break-inside: avoid;\n        position: relative;\n        border: 2px solid transparent;\n      }\n  \n      #preview-content h5::after, #setText h5::after {\n        content: \"\";\n        display: block;\n        height: 100px;\n        margin-bottom: -100px;\n        position: relative;\n      }\n  \n      #preview-content h6, #setText h6 {\n        page-break-inside: avoid;\n        position: relative;\n        border: 2px solid transparent;\n      }\n  \n      #preview-content h6::after, #setText h6::after {\n        content: \"\";\n        display: block;\n        height: 100px;\n        margin-bottom: -100px;\n        position: relative;\n      }\n    }\n    #preview-content sup, #setText sup {\n      top: -.5em;\n      position: relative;\n      font-size: 75%;\n      line-height: 0;\n      vertical-align: baseline;\n    }\n    \n    #preview-content .text-url, #setText .text-url {\n      " + (useColors ? 'color: #0B93ff;' : '') + "\n      cursor: text;\n      pointer-events: none;\n    }\n    \n    #preview-content .text-url a:hover, #setText .text-url a:hover {\n      " + (useColors ? 'color: #0B93ff;' : '') + "\n    }\n";
+};
+exports.PreviewStyle = "\n  #preview {\n    font-family: 'CMU Serif', 'Georgia', Helvetica, Arial, sans-serif;\n    font-size: 17px;\n    visibility: visible;\n    word-break: break-word;\n    padding: 2.5em;\n    max-width: 800px;\n    margin: auto;\n    box-sizing: content-box;\n  }\n\n  #preview h1, #preview h2, #preview h3, #preview h4, #preview h5, #preview strong {\n    font-family: 'CMU Serif Bold', 'Georgia', Helvetica, Arial, sans-serif;\n  }\n\n  #preview  i, #preview  em {\n    font-family: 'CMU Serif Italic', 'Georgia', Helvetica, Arial, sans-serif;\n  }\n";
+exports.TocStyle = function (containerName) {
+    if (containerName === void 0) { containerName = 'toc'; }
+    return "\n  #" + containerName + " {\n    width: 25%;\n    overflow-y: auto;\n    min-width: 370px; \n  }\n  @media only screen and (max-width: 960px) {\n    #{containerName}  {\n      display: none; \n    } \n  }\n  @media print {\n    #{containerName}  {\n      display: none; \n    } \n  } \n  #toc_container {\n    padding: 70px 0 0 60px; \n  }\n  #toc_container .toc-title-1 {\n    font-weight: 500; \n  }\n  #toc_container .toc-title-1 > a {\n    font-size: 1.6em; \n  }\n  #toc_container .toc-title-2 > a {\n    font-size: 1.3em; \n  }\n  #toc_container .toc-title-3,  #toc_container .toc-title-4,  #toc_container .toc-title-5, .previewPage #toc_container .toc-title-6 {\n    padding-left: 20px; \n  }\n  #toc_container .toc-title-3 > a,  #toc_container .toc-title-4 > a,  #toc_container .toc-title-5 > a, .previewPage #toc_container .toc-title-6 > a {\n    font-size: 1em;\n    color: #979797; \n  }\n  #toc_container ul {\n    list-style: none;\n    padding: 0;\n    margin: 0; \n  }\n  #toc_container li {\n    padding-top: 20px; \n  }\n  #toc_container li a {\n    color: #000; \n  }\n  #toc_container li a:hover {\n    color: #0093FF; \n  }\n  #toc_container li a:active,\n  #toc_container li .selected {\n    color: #047DD6 !important; \n  }\n";
+};
+//# sourceMappingURL=index.js.map
