+++ lib/styles/styles-container.d.ts
@@ -1,0 +1,1 @@
+export declare const ContainerStyle: (useColors?: boolean) => string;
