+++ lib/styles/styles-container.js
@@ -1,0 +1,8 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.ContainerStyle = void 0;
+exports.ContainerStyle = function (useColors) {
+    if (useColors === void 0) { useColors = true; }
+    return "\n  html,body {\n    width: 100%;\n    height: 100%;\n  }\n  *, *::before,*::after {\n    box-sizing: border-box;\n  }\n  @-ms-viewport {\n    width: device-width;\n  }\n  body {\n    margin: 0;\n    " + (useColors ? 'color: #1E2029;' : '') + "\n    font-size: 14px;\n    line-height: normal;\n  }\n  hr {\n    box-sizing: content-box;\n    height: 0;\n    overflow: visible;\n  }\n  h1, h2, h3, h4, h5, h6 {\n    margin-top: 0;\n    margin-bottom: 0.5em;\n    " + (useColors ? 'color: rgba(0, 0, 0, 0.85);' : '') + "\n    font-weight: 500;\n  }\n  p {\n    margin-top: 0;\n    margin-bottom: 1em;\n  }\n  ol, ul, dl {\n    margin-top: 0;\n    margin-bottom: 1em;\n  }\n  ol ol, ul ul, ol ul, ul ol {\n    margin-bottom: 0;\n  }\n  dt {\n    font-weight: 500;\n  }\n  dd {\n    margin-bottom: 0.5em;\n    margin-left: 0;\n  }\n  blockquote {\n    margin: 0 0 1em;\n  }\n  dfn {\n    font-style: italic;\n  }\n  b, strong {\n    font-weight: bolder;\n  }\n  small {\n    font-size: 80%;\n  }\n  sub, sup {\n    position: relative;\n    font-size: 75%;\n    line-height: 0;\n    vertical-align: baseline;\n  }\n  sub {\n    bottom: -0.25em;\n  }\n  sup {\n    top: -0.5em;\n  }\n  a {\n    " + (useColors ? 'color: #0B93ff;' : '') + "\n    text-decoration: none;\n    background-color: transparent;\n    outline: none;\n    cursor: pointer;\n    transition: color 0.3s;\n  }\n  a:hover {\n    " + (useColors ? 'color: #33aaff;' : '') + "\n  }\n  a:active {\n    " + (useColors ? 'color: #0070d9;' : '') + "\n  }\n  a:active, a:hover {\n    text-decoration: none;\n    outline: 0;\n  }\n  a[disabled] {\n    " + (useColors ? 'color: rgba(0, 0, 0, 0.25)' : '') + "\n    cursor: not-allowed;\n    pointer-events: none;\n  }\n  pre, code, kbd, samp {\n    font-size: 1em;\n  }\n  pre {\n    margin-top: 0;\n    margin-bottom: 1em;\n    overflow: auto;\n  }\n  figure {\n    margin: 0 0 1em;\n  }\n  img {\n    vertical-align: middle;\n    border-style: none;\n  }\n  svg:not(:root) {\n    overflow: hidden;\n  }\n  table {\n    border-collapse: collapse;\n  }\n  caption {\n    padding-top: 0.75em;\n    padding-bottom: 0.3em;\n    " + (useColors ? 'color: rgba(0, 0, 0, 0.45)' : '') + "\n    text-align: left;\n    caption-side: bottom;\n  }\n  th {\n    text-align: inherit;\n  }\n";
+};
+//# sourceMappingURL=styles-container.js.map
