+++ lib/styles/styles-fonts.d.ts
@@ -1,0 +1,1 @@
+export declare const fontsStyles = "\n        @import url('https://fonts.googleapis.com/css?family=IBM+Plex+Sans:400,500');\n        @import url('https://fonts.googleapis.com/css?family=Source+Sans+Pro:400,700');\n        @import url('https://fonts.googleapis.com/css?family=Inconsolata');\n        @import url('https://cdn.mathpix.com/fonts/cmu.css');\n";
