+++ lib/styles/styles-fonts.js
@@ -1,0 +1,5 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.fontsStyles = void 0;
+exports.fontsStyles = "\n        @import url('https://fonts.googleapis.com/css?family=IBM+Plex+Sans:400,500');\n        @import url('https://fonts.googleapis.com/css?family=Source+Sans+Pro:400,700');\n        @import url('https://fonts.googleapis.com/css?family=Inconsolata');\n        @import url('https://cdn.mathpix.com/fonts/cmu.css');\n";
+//# sourceMappingURL=styles-fonts.js.map
