+++ lib/styles/styles-lists.d.ts
@@ -1,0 +1,1 @@
+export declare const listsStyles = "\n  ol.enumerate, ul.itemize {\n    padding-inline-start: 40px;\n  }\n/* It's commented because counter not supporting to change value \n  ol.enumerate.lower-alpha {\n    counter-reset: item ;\n    list-style-type: none !important;\n  }\n  .enumerate.lower-alpha > li {\n    position: relative;\n  }\n  .enumerate.lower-alpha > li:before { \n    content: \"(\"counter(item, lower-alpha)\")\"; \n    counter-increment: item; \n    position: absolute;\n    left: -47px;\n    width: 47px;\n    display: flex;\n    justify-content: flex-end;\n    padding-right: 7px;\n    flex-wrap: nowrap;\n    word-break: keep-all;\n  }\n  */\n  \n  .itemize > li {\n    position: relative;\n  }\n  .itemize > li > span.li_level { \n    position: absolute;\n    left: -40px;\n    width: 40px;\n    display: flex;\n    justify-content: flex-end;\n    padding-right: 10px;\n    box-sizing: border-box;\n  }\n";
