+++ lib/styles/styles-tabular.d.ts
@@ -1,0 +1,1 @@
+export declare const tabularStyles: (useColors?: boolean) => string;
