+++ lib/styles/styles-tabular.js
@@ -1,0 +1,8 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.tabularStyles = void 0;
+exports.tabularStyles = function (useColors) {
+    if (useColors === void 0) { useColors = true; }
+    return "\n    .table_tabular table th,  .table_tabular table th {\n        border: none !important;\n        padding: 6px 13px;\n    }\n      \n    #tabular tr, #tabular tr {\n        border-top: none !important;\n        border-bottom: none !important;\n    }\n    #tabular td, #tabular td {\n        border-style: none !important;\n        " + (useColors ? 'background-color: #fff;' : 'background-color: transparent;') + "\n        " + (useColors ? 'border-color: #000 !important;' : 'border-color: currentColor !important;') + "\n        word-break: keep-all;\n        padding: 0.1em 0.5em !important;\n    }\n    #tabular {\n        display: inline !important;\n    }\n    #tabular td > p {\n        margin-bottom: 0;\n        margin-top: 0;\n    }\n    #tabular td._empty {\n      height: 1.3em;\n    }\n    #tabular td .f {\n      opacity: 0;\n    }\n    \n    html[data-theme=\"dark\"] #tabular tr, html[data-theme=\"dark\"] #tabular td {\n      " + (useColors ? 'background-color: #202226;' : 'background-color: transparent;') + "\n      " + (useColors ? 'border-color: #fff !important;' : 'border-color: currentColor !important;') + "\n    }  \n    .table_tabular {\n        overflow-x: auto;\n        padding: 0 2px 0.5em 2px;\n    }\n    .figure_img {\n       margin-bottom: 0.5em;\n       overflow-x: auto;\n    }\n";
+};
+//# sourceMappingURL=styles-tabular.js.map
