--- src/mathjax/styles-lists.ts
@@ -1,39 +1,0 @@
-export const listsStyles = `
-  ol.enumerate, ul.itemize {
-    padding-inline-start: 40px;
-  }
-/* It's commented because counter not supporting to change value 
-  ol.enumerate.lower-alpha {
-    counter-reset: item ;
-    list-style-type: none !important;
-  }
-  .enumerate.lower-alpha > li {
-    position: relative;
-  }
-  .enumerate.lower-alpha > li:before { 
-    content: "("counter(item, lower-alpha)")"; 
-    counter-increment: item; 
-    position: absolute;
-    left: -47px;
-    width: 47px;
-    display: flex;
-    justify-content: flex-end;
-    padding-right: 7px;
-    flex-wrap: nowrap;
-    word-break: keep-all;
-  }
-  */
-  
-  .itemize > li {
-    position: relative;
-  }
-  .itemize > li > span.li_level { 
-    position: absolute;
-    left: -40px;
-    width: 40px;
-    display: flex;
-    justify-content: flex-end;
-    padding-right: 10px;
-    box-sizing: border-box;
-  }
-`;
