+++ src/styles/styles-tabular.ts
@@ -1,0 +1,44 @@
+export const tabularStyles = (useColors = true) => `
+    .table_tabular table th,  .table_tabular table th {
+        border: none !important;
+        padding: 6px 13px;
+    }
+      
+    #tabular tr, #tabular tr {
+        border-top: none !important;
+        border-bottom: none !important;
+    }
+    #tabular td, #tabular td {
+        border-style: none !important;
+        ${useColors ? 'background-color: #fff;' : 'background-color: transparent;'}
+        ${useColors ? 'border-color: #000 !important;' : 'border-color: currentColor !important;'}
+        word-break: keep-all;
+        padding: 0.1em 0.5em !important;
+    }
+    #tabular {
+        display: inline !important;
+    }
+    #tabular td > p {
+        margin-bottom: 0;
+        margin-top: 0;
+    }
+    #tabular td._empty {
+      height: 1.3em;
+    }
+    #tabular td .f {
+      opacity: 0;
+    }
+    
+    html[data-theme="dark"] #tabular tr, html[data-theme="dark"] #tabular td {
+      ${useColors ? 'background-color: #202226;' : 'background-color: transparent;'}
+      ${useColors ? 'border-color: #fff !important;' : 'border-color: currentColor !important;'}
+    }  
+    .table_tabular {
+        overflow-x: auto;
+        padding: 0 2px 0.5em 2px;
+    }
+    .figure_img {
+       margin-bottom: 0.5em;
+       overflow-x: auto;
+    }
+`;
